refactor(deblock): move triage out of library, keep as harness-only research
Full-corpus evaluation showed triage can't beat bilateral_bnd or
boundary_4tap at any quality level. No reason to expose it as a
library-level deblock strategy.

- Removed zenjpeg/src/deblock/triage.rs and its re-exports from
  deblock/mod.rs.
- Inlined the filter as a private `mod triage_impl` inside
  deblock_harness.rs so TriageDecode / TriageGated continue to run
  as academic-interest comparison strategies in the harness.

No quality or performance change; output bit-identical to pre-move.
